Flask Update
this update lets the user use the viewing shopping cart feature on flask and postman to PUT,POST,GET, and DELETE

GET /cart: Retrieves the current items in the shopping cart.
POST /cart: Adds a book to the shopping cart. If the book is already in the cart, it updates the quantity.
PUT /cart/<title>: Updates the quantity of a specific book in the shopping cart.
DELETE /cart/<title>: Removes a specific book from the shopping cart.